Implement ingestion logic and models
