+++ newBackend/.env
@@ -3,3 +3,10 @@
 
 #Load path for dataset here
 PROCESSED_DATA_PATH=./mock_data/processed_data.json
+
+#for pool.js
+DB_USER=postgres
+DB_PASSWORD=postgrespw
+DB_HOST=localhost
+DB_PORT=5432
+DB_NAME=IoTDatabase
