+++ newBackend/src/db/pool.js
@@ -1,0 +1,13 @@
+require('dotenv').config({ path: '../.env' });
+
+const { Pool } = require('pg');
+
+const pool = new Pool({
+  user: process.env.DB_USER,
+  host: process.env.DB_HOST,
+  database: process.env.DB_NAME,
+  password: process.env.DB_PASSWORD,
+  port: process.env.DB_PORT
+});
+
+module.exports = pool;
