+++ newBackend/src/models/dataset.js
@@ -1,0 +1,8 @@
+class Dataset {
+  constructor({ id, name }) {
+    this.id = id;
+    this.name = name;
+  }
+}
+
+module.exports = Dataset;
