+++ newBackend/src/models/timeSeriesPoint.js
@@ -1,0 +1,21 @@
+class TimeSeriesPoint {
+  constructor({
+    id,
+    datasetId,
+    entity,
+    metric,
+    ts,
+    value,
+    qualityFlag
+  }) {
+    this.id = id;
+    this.datasetId = datasetId;
+    this.entity = entity;
+    this.metric = metric;
+    this.ts = ts;
+    this.value = value;
+    this.qualityFlag = qualityFlag;
+  }
+}
+
+module.exports = TimeSeriesPoint;
